pyproject.toml
[tool.poetry]
name = "lakehouse"
version = "0.1.0"
description = ""
authors = ["Robert Pack <robstar.pack@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10,<=3.11"
dagster = "^1.5.9"
dagster-webserver = "^1.5.9"
dagster-deltalake = "^0.21.9"
dagster-deltalake-pandas = "^0.21.9"
scikit-learn = "^1.3"
scipy = "^1.9"
object-store-python = "^0.1.10"

[tool.poetry.group.dev.dependencies]
requests = "^2.31.0"
ipykernel = "^6.25.1"
pyspark = "^3.5.0"
delta-spark = "^3.0.0"
polars = ">=0.19.13"
ruff = "^0.1.6"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
